Move: Products에서 Search로 관련 컴포넌트 이동
