+++ src/pages/search/components/BottomActions/styles.ts
