+++ src/pages/search/components/SearchTab/styles.ts
